A new file added
In Time Series Forecasting
Aim :- Using the past data related to predict the future data points.

Assumption:- Past behaviour will have an influence on the future behaviour.

ARIMA model
ARIMA is a model which is used for time series forecasting. There are 3 major components in ARIMA.

1) AR - Auto Regressive Component

2) I - Integerated ( How many time we are differencing to make the data stationary.

3) MA - Moving Average Component

A time series data using ARIMA model can be forecasted in two ways. Regression and Moving Average.

Regression - It is forecating the value based on previous input values.

        Y = mX + C + e
Moving Average - It is forecasting the values based on previous residuals ( Error).